tests/Element-setAttribute.html

<!DOCTYPE html>
<title>Element.setAttribute</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="../web-dom-core#dom-Element-setAttribute">
<script src="testharness.js"></script>
<div id="log"></div>
<script>
// XXX some of this is testing Attr and getAttribute / getAttributeNS
// maybe it should all be merged?
test(function() {
  var el = document.createElement("foo"),
      invalid = ["\\", "'", '"', "0", "0:a"],
      valid = ["x", ":", "a:0"]
  for(var i = 0; i < invalid.length; i++) {
    assert_throws("INVALID_CHARACTER_ERR", function() { el.setAttribute(invalid[i], "test") })
  }
  for(var i = 0; i < valid.length; i++) {
    el.setAttribute(valid[i], "test")
    assert_equals(el.getAttribute(valid[i]), "test")
  }

  el = document.createElement("bar")
  el.setAttribute("a", "1")
  el.setAttribute("b", "2")
  el.setAttribute("a", "3")
  el.setAttribute("c", "4")
  var attrs = el.attributes
  var expected = [{"n":"a","v":"3"}, {"n":"b","v":"2"}, {"n":"c","v":"4"}]
  for (var i = 0, il = attrs.length; i < il; i++) {
    assert_equals(attrs[i].name, expected[i].n)
    assert_equals(attrs[i].value, expected[i].v)
    assert_equals(attrs[i].specified, true)
    assert_equals(attrs[i].namespaceURI, null)
    assert_equals(attrs[i].ownerElement, el)
  }

  el = document.createElement("baz")
  el.setAttributeNS("ab", "attr", "fail")
  el.setAttributeNS("kl", "attr", "pass")
  el.setAttribute("attr", "pass")
  attrs = el.attributes
  for (var i = 0, il = attrs.length; i < il; i++) {
    assert_equals(attrs[i].value, "pass")
  }

  el = document.createElement("iframe")
  el.setAttribute("src", "file:///home")
  assert_equals(el.getAttribute("src"), "file:///home")

  el = document.createElement("div")
  el.setAttribute("ALIGN", "left")
  assert_equals(el.getAttributeNS("", "ALIGN"), "")
  assert_equals(el.getAttributeNS("", "align"), "left")
  assert_equals(el.getAttribute("align"), "left")
  el.setAttribute("CHEEseCaKe", "tasty")
  assert_equals(el.getAttributeNS("", "CHEEseCaKe"), "")
  assert_equals(el.getAttributeNS("", "cheesecake"), "tasty")
  assert_equals(el.getAttribute("cheesecake"), "tasty")
})
</script>
